# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StockBTP - API de gestion de stock pour entreprise de construction de conteneurs
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductUnit = typeof ProductUnit[keyof typeof ProductUnit];


export const ProductUnit = {
  kg: 'kg',
  m: 'm',
  litre: 'litre',
  piece: 'piece',
} as const;

export type ProductLocation = typeof ProductLocation[keyof typeof ProductLocation];


export const ProductLocation = {
  warehouse: 'warehouse',
  site: 'site',
  project: 'project',
} as const;

export interface Product {
  id: number;
  name: string;
  category: string;
  unit: ProductUnit;
  quantityInStock: number;
  minimumThreshold: number;
  location: ProductLocation;
  createdAt: string;
  updatedAt: string;
}

export type ProductInputUnit = typeof ProductInputUnit[keyof typeof ProductInputUnit];


export const ProductInputUnit = {
  kg: 'kg',
  m: 'm',
  litre: 'litre',
  piece: 'piece',
} as const;

export type ProductInputLocation = typeof ProductInputLocation[keyof typeof ProductInputLocation];


export const ProductInputLocation = {
  warehouse: 'warehouse',
  site: 'site',
  project: 'project',
} as const;

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  unit: ProductInputUnit;
  quantityInStock?: number;
  minimumThreshold: number;
  location: ProductInputLocation;
}

export type ProductUpdateUnit = typeof ProductUpdateUnit[keyof typeof ProductUpdateUnit];


export const ProductUpdateUnit = {
  kg: 'kg',
  m: 'm',
  litre: 'litre',
  piece: 'piece',
} as const;

export type ProductUpdateLocation = typeof ProductUpdateLocation[keyof typeof ProductUpdateLocation];


export const ProductUpdateLocation = {
  warehouse: 'warehouse',
  site: 'site',
  project: 'project',
} as const;

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  unit?: ProductUpdateUnit;
  minimumThreshold?: number;
  location?: ProductUpdateLocation;
}

export interface CategoryStat {
  category: string;
  totalQuantity: number;
  productCount: number;
}

export type StockMovementType = typeof StockMovementType[keyof typeof StockMovementType];


export const StockMovementType = {
  IN: 'IN',
  OUT: 'OUT',
} as const;

export interface StockMovement {
  id: number;
  productId: number;
  /** @nullable */
  productName?: string | null;
  type: StockMovementType;
  quantity: number;
  reason: string;
  /** @nullable */
  projectId?: number | null;
  /** @nullable */
  projectName?: string | null;
  createdById: number;
  /** @nullable */
  createdByName?: string | null;
  createdAt: string;
}

export type StockMovementInputType = typeof StockMovementInputType[keyof typeof StockMovementInputType];


export const StockMovementInputType = {
  IN: 'IN',
  OUT: 'OUT',
} as const;

export interface StockMovementInput {
  productId: number;
  type: StockMovementInputType;
  /** @minimum 1 */
  quantity: number;
  /** @minLength 1 */
  reason: string;
  /** @nullable */
  projectId?: number | null;
  createdById: number;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  clientName?: string | null;
  status: ProjectStatus;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  createdAt: string;
}

export type ProjectInputStatus = typeof ProjectInputStatus[keyof typeof ProjectInputStatus];


export const ProjectInputStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  clientName?: string;
  status: ProjectInputStatus;
  startDate?: string;
  endDate?: string;
}

export type ProjectUpdateStatus = typeof ProjectUpdateStatus[keyof typeof ProjectUpdateStatus];


export const ProjectUpdateStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  clientName?: string;
  status?: ProjectUpdateStatus;
  startDate?: string;
  endDate?: string;
}

export interface ProjectMaterial {
  id: number;
  projectId: number;
  productId: number;
  /** @nullable */
  productName?: string | null;
  quantityUsed: number;
  /** @nullable */
  unit?: string | null;
}

export interface ProjectMaterialInput {
  productId: number;
  /** @minimum 1 */
  quantityUsed: number;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  manager: 'manager',
  worker: 'worker',
} as const;

export interface User {
  id: number;
  fullName: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  manager: 'manager',
  worker: 'worker',
} as const;

export interface UserInput {
  /** @minLength 1 */
  fullName: string;
  email: string;
  role: UserInputRole;
  /** @minLength 6 */
  password: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  manager: 'manager',
  worker: 'worker',
} as const;

export interface UserUpdate {
  fullName?: string;
  role?: UserUpdateRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface DashboardSummary {
  totalProducts: number;
  totalStockValue: number;
  lowStockCount: number;
  activeProjects: number;
  todayMovementsIn: number;
  todayMovementsOut: number;
}

export type ListProductsParams = {
category?: string;
low_stock?: boolean;
search?: string;
location?: string;
limit?: number;
offset?: number;
};

export type ListStockMovementsParams = {
product_id?: number;
project_id?: number;
type?: ListStockMovementsType;
from_date?: string;
to_date?: string;
limit?: number;
offset?: number;
};

export type ListStockMovementsType = typeof ListStockMovementsType[keyof typeof ListStockMovementsType];


export const ListStockMovementsType = {
  IN: 'IN',
  OUT: 'OUT',
} as const;

export type ListProjectsParams = {
status?: ListProjectsStatus;
};

export type ListProjectsStatus = typeof ListProjectsStatus[keyof typeof ListProjectsStatus];


export const ListProjectsStatus = {
  active: 'active',
  completed: 'completed',
  paused: 'paused',
} as const;

export type GetRecentMovementsParams = {
limit?: number;
};