# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: StockBTP - API de gestion de stock pour entreprise de construction de conteneurs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: Liste tous les produits
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: low_stock
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Liste des produits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Créer un produit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Produit créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Détails d'un produit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Produit trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Produit non trouvé
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Modifier un produit
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Produit modifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Supprimer un produit (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Supprimé
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: Liste des projets
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [active, completed, paused] }
      responses:
        "200":
          description: Liste des projets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Créer un projet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Projet créé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Détails d'un projet
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Projet trouvé
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Non trouvé
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Modifier un projet
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Projet modifié
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/materials:
    get:
      operationId: getProjectMaterials
      tags: [project-materials]
      summary: Matériaux utilisés sur un projet
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Matériaux du projet
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMaterial"
    post:
      operationId: addProjectMaterial
      tags: [project-materials]
      summary: Ajouter un matériau à un projet
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectMaterialInput"
      responses:
        "201":
          description: Matériau ajouté
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMaterial"

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /auth/login:
    post:
      operationId: login
      tags: [users]
      summary: Connexion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Identifiants invalides
...
  /auth/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Utilisateur courant
      responses:
        "200":
          description: Utilisateur courant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Non authentifié

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, category, unit, quantityInStock, minimumThreshold, location, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        unit: { type: string, enum: [kg, m, litre, piece] }
        quantityInStock: { type: integer }
        minimumThreshold: { type: integer }
        location: { type: string, enum: [warehouse, site, project] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProductInput:
      type: object
      required: [name, category, unit, minimumThreshold, location]
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        unit: { type: string, enum: [kg, m, litre, piece] }
        quantityInStock: { type: integer, default: 0 }
        minimumThreshold: { type: integer, default: 0 }
        location: { type: string, enum: [warehouse, site, project] }
    ProductUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        category: { type: string }
        unit: { type: string, enum: [kg, m, litre, piece] }
        minimumThreshold: { type: integer }
        location: { type: string, enum: [warehouse, site, project] }
...
    Project:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        clientName: { type: ["string", "null"] }
        status: { type: string, enum: [active, completed, paused] }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        createdAt: { type: string }
    ProjectInput:
      type: object
      required: [name, status]
      properties:
        name: { type: string, minLength: 1 }
        clientName: { type: string }
        status: { type: string, enum: [active, completed, paused] }
        startDate: { type: string }
        endDate: { type: string }
    ProjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        clientName: { type: string }
        status: { type: string, enum: [active, completed, paused] }
        startDate: { type: string }
        endDate: { type: string }

    # ─── PROJECT MATERIAL ────────────────────────────────────────────────────
    ProjectMaterial:
      type: object
      required: [id, projectId, productId, quantityUsed]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        productId: { type: integer }
        productName: { type: ["string", "null"] }
        quantityUsed: { type: integer }
        unit: { type: ["string", "null"] }
    ProjectMaterialInput:
      type: object
      required: [productId, quantityUsed]
      properties:
        productId: { type: integer }
        quantityUsed: { type: integer, minimum: 1 }

    # ─── USER ────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, fullName, email, role, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string }
        role: { type: string, enum: [admin, manager, worker] }
        createdAt: { type: string }
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token: { type: string }

    # ─── DASHBOARD ──────────────────────────────────────────────────────────
...