# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StockBTP - API de gestion de stock pour entreprise de construction de conteneurs
 * OpenAPI spec version: 0.1.0
 */
import type { ProductLocation } from './productLocation';
import type { ProductUnit } from './productUnit';

export interface Product {
  id: number;
  name: string;
  category: string;
  unit: ProductUnit;
  quantityInStock: number;
  minimumThreshold: number;
  location: ProductLocation;
  createdAt: string;
  updatedAt: string;
}