# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * StockBTP - API de gestion de stock pour entreprise de construction de conteneurs
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputLocation } from './productInputLocation';
import type { ProductInputUnit } from './productInputUnit';

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  unit: ProductInputUnit;
  quantityInStock?: number;
  minimumThreshold: number;
  location: ProductInputLocation;
}